Change types to silence new compiler warnings.

+++ src/stan/agrad/rev/matrix/mdivide_left_ldlt.hpp
@@ -41,10 +41,10 @@ namespace stan {
           _alloc(new mdivide_left_ldlt_alloc<R1,C1,R2,C2>()),
           _alloc_ldlt(A._alloc)
         {
-          size_t pos = 0;
+          int pos = 0;
           _alloc->C_.resize(M_,N_);
-          for (size_t j = 0; j < N_; j++) {
-            for (size_t i = 0; i < M_; i++) {
+          for (int j = 0; j < N_; j++) {
+            for (int i = 0; i < M_; i++) {
               _variRefB[pos] = B(i,j).vi_;
               _alloc->C_(i,j) = B(i,j).val();
               pos++;
@@ -54,8 +54,8 @@ namespace stan {
           _alloc_ldlt->_ldlt.solveInPlace(_alloc->C_);
           
           pos = 0;
-          for (size_t j = 0; j < N_; j++) {
-            for (size_t i = 0; i < M_; i++) {
+          for (int j = 0; j < N_; j++) {
+            for (int i = 0; i < M_; i++) {
               _variRefC[pos] = new vari(_alloc->C_(i,j),false);
               pos++;
             }
@@ -66,21 +66,21 @@ namespace stan {
           Eigen::Matrix<double,R1,C1> adjA(M_,M_);
           Eigen::Matrix<double,R2,C2> adjB(M_,N_);
           
-          size_t pos = 0;
-          for (size_t j = 0; j < N_; j++)
-            for (size_t i = 0; i < M_; i++)
+          int pos = 0;
+          for (int j = 0; j < N_; j++)
+            for (int i = 0; i < M_; i++)
               adjB(i,j) = _variRefC[pos++]->adj_;
           
           _alloc_ldlt->_ldlt.solveInPlace(adjB);
           adjA.noalias() = -adjB * _alloc->C_.transpose();
 
-          for (size_t j = 0; j < M_; j++)
-            for (size_t i = 0; i < M_; i++)
+          for (int j = 0; j < M_; j++)
+            for (int i = 0; i < M_; i++)
               _alloc_ldlt->_variA(i,j)->adj_ += adjA(i,j);
           
           pos = 0;
-          for (size_t j = 0; j < N_; j++)
-            for (size_t i = 0; i < M_; i++)
+          for (int j = 0; j < N_; j++)
+            for (int i = 0; i < M_; i++)
               _variRefB[pos++]->adj_ += adjB(i,j);
         }
       };
@@ -108,10 +108,10 @@ namespace stan {
           using Eigen::Matrix;
           using Eigen::Map;
           
-          size_t pos = 0;
+          int pos = 0;
           _alloc->C_.resize(M_,N_);
-          for (size_t j = 0; j < N_; j++) {
-            for (size_t i = 0; i < M_; i++) {
+          for (int j = 0; j < N_; j++) {
+            for (int i = 0; i < M_; i++) {
               _variRefB[pos] = B(i,j).vi_;
               _alloc->C_(i,j) = B(i,j).val();
               pos++;
@@ -122,8 +122,8 @@ namespace stan {
           _alloc->_ldltP->solveInPlace(_alloc->C_);
           
           pos = 0;
-          for (size_t j = 0; j < N_; j++) {
-            for (size_t i = 0; i < M_; i++) {
+          for (int j = 0; j < N_; j++) {
+            for (int i = 0; i < M_; i++) {
               _variRefC[pos] = new vari(_alloc->C_(i,j),false);
               pos++;
             }
@@ -133,16 +133,16 @@ namespace stan {
         virtual void chain() {
           Eigen::Matrix<double,R2,C2> adjB(M_,N_);
           
-          size_t pos = 0;
-          for (size_t j = 0; j < adjB.cols(); j++)
-            for (size_t i = 0; i < adjB.rows(); i++)
+          int pos = 0;
+          for (int j = 0; j < adjB.cols(); j++)
+            for (int i = 0; i < adjB.rows(); i++)
               adjB(i,j) = _variRefC[pos++]->adj_;
           
           _alloc->_ldltP->solveInPlace(adjB);
           
           pos = 0;
-          for (size_t j = 0; j < adjB.cols(); j++)
-            for (size_t i = 0; i < adjB.rows(); i++)
+          for (int j = 0; j < adjB.cols(); j++)
+            for (int i = 0; i < adjB.rows(); i++)
               _variRefB[pos++]->adj_ += adjB(i,j);
         }
       };
@@ -169,9 +169,9 @@ namespace stan {
           _alloc->C_ = B;
           _alloc_ldlt->_ldlt.solveInPlace(_alloc->C_);
           
-          size_t pos = 0;
-          for (size_t j = 0; j < N_; j++) {
-            for (size_t i = 0; i < M_; i++) {
+          int pos = 0;
+          for (int j = 0; j < N_; j++) {
+            for (int i = 0; i < M_; i++) {
               _variRefC[pos] = new vari(_alloc->C_(i,j),false);
               pos++;
             }
@@ -182,15 +182,15 @@ namespace stan {
           Eigen::Matrix<double,R1,C1> adjA(M_,M_);
           Eigen::Matrix<double,R1,C2> adjC(M_,N_);
 
-          size_t pos = 0;
-          for (size_t j = 0; j < adjC.cols(); j++)
-            for (size_t i = 0; i < adjC.rows(); i++)
+          int pos = 0;
+          for (int j = 0; j < adjC.cols(); j++)
+            for (int i = 0; i < adjC.rows(); i++)
               adjC(i,j) = _variRefC[pos++]->adj_;
         
           adjA = -_alloc_ldlt->_ldlt.solve(adjC*_alloc->C_.transpose());
 
-          for (size_t j = 0; j < adjA.cols(); j++)
-            for (size_t i = 0; i < adjA.rows(); i++)
+          for (int j = 0; j < adjA.cols(); j++)
+            for (int i = 0; i < adjA.rows(); i++)
               _alloc_ldlt->_variA(i,j)->adj_ += adjA(i,j);
         }
       };
@@ -214,9 +214,9 @@ namespace stan {
       
       mdivide_left_ldlt_vv_vari<R1,C1,R2,C2> *baseVari = new mdivide_left_ldlt_vv_vari<R1,C1,R2,C2>(A,b);
       
-      size_t pos = 0;
-      for (size_t j = 0; j < res.cols(); j++)
-        for (size_t i = 0; i < res.rows(); i++)
+      int pos = 0;
+      for (int j = 0; j < res.cols(); j++)
+        for (int i = 0; i < res.rows(); i++)
           res(i,j).vi_ = baseVari->_variRefC[pos++];
       
       return res;
@@ -232,9 +232,9 @@ namespace stan {
       
       mdivide_left_ldlt_vd_vari<R1,C1,R2,C2> *baseVari = new mdivide_left_ldlt_vd_vari<R1,C1,R2,C2>(A,b);
       
-      size_t pos = 0;
-      for (size_t j = 0; j < res.cols(); j++)
-        for (size_t i = 0; i < res.rows(); i++)
+      int pos = 0;
+      for (int j = 0; j < res.cols(); j++)
+        for (int i = 0; i < res.rows(); i++)
           res(i,j).vi_ = baseVari->_variRefC[pos++];
       
       return res;
@@ -250,9 +250,9 @@ namespace stan {
       
       mdivide_left_ldlt_dv_vari<R1,C1,R2,C2> *baseVari = new mdivide_left_ldlt_dv_vari<R1,C1,R2,C2>(A,b);
       
-      size_t pos = 0;
-      for (size_t j = 0; j < res.cols(); j++)
-        for (size_t i = 0; i < res.rows(); i++)
+      int pos = 0;
+      for (int j = 0; j < res.cols(); j++)
+        for (int i = 0; i < res.rows(); i++)
           res(i,j).vi_ = baseVari->_variRefC[pos++];
       
       return res;
